+++ LustreCollector/FileSystem/IFileSystemChangeWatcher.cs
@@ -2,5 +2,5 @@ namespace LustreCollector.FileSystem;
 
 public interface IFileSystemChangeWatcher
 {
-    public IAsyncEnumerable<FilesystemChangeEvent> Watch(DirectoryInfo root);
+    public IAsyncEnumerable<FileSystemChangeEvent> Watch(DirectoryInfo root);
 }
